+++ assets/tpls/preload.php
@@ -1,0 +1,3 @@
+<?php
+
+define('DOKU_CONF', '/data/conf/');